==> gaingrn/notebooks/1_template_selection.ipynb <==
# 1 Template Selection
A Notebook for Picking suitable Template Structures from Subgroups of the GAIN domain Dataset.


*To run this notebook, you will need to download all PDB models and the GAIN-GRN data from the zenodo repository:* 

https://dx.doi.org/10.5281/zenodo.12515545/gaingrn_data.tgz : download via `gaingrn.utils.io.download_data()` into `path/to/GAIN-GRN/data`

https://dx.doi.org/10.5281/zenodo.12515545/agpcr_gains.tgz : download via `gaingrn.utils.io.download_pdbs(target_directory=my/dir/to/pdbs)` and specify the `PDB_DIR` variable.

# DEPENDENCIES
import glob, glob, re, os, sys
#from shutil import copyfile
import numpy as np
import pandas as pd
# LOCAL IMPORTS
import gaingrn.utils.gain_classes
import gaingrn.utils.template_utils
import gaingrn.utils.assign
import gaingrn.utils.plotting_utils
# FIX FOR CHANGED PICKLE PATHING
sys.modules['gaingrn.scripts.gain_classes'] = gaingrn.utils.gain_classes

try: 
    GESAMT_BIN = os.environ.get('GESAMT_BIN')
except:
    GESAMT_BIN = "/home/hildilab/lib/xtal/ccp4-8.0/ccp4-8.0/bin/gesamt"

if GESAMT_BIN is None:
    GESAMT_BIN = "/home/hildilab/lib/xtal/ccp4-8.0/ccp4-8.0/bin/gesamt"

PDB_DIR = "../../all_pdbs"

# This is a remnant from the previous approach in having MSA-based anchors and is used as a comparative metric.
valid_seqs = gaingrn.utils.io.read_multi_seq("../data/seq_aln/offset_valid_seqs.fa")
stride_files = glob.glob("../data/gain_strides/*.stride")

### 1. Load the full GainCollection of the Dataset and habe the folder containing all GAIN PDB files.

valid_collection = pd.read_pickle("../data/valid_collection.pkl")
allpdbs = glob.glob(f'{PDB_DIR}/*.pdb')
print(len(allpdbs, "PDB files have been found."))

### 2.Chunk down the Dataset into Subselections containing every subfamily ("A", "B") and receptor type ("B1", "D2")

fam_list = [gaingrn.utils.io.get_agpcr_type(gain.name) for gain in valid_collection.collection]
name_list = [gain.name for gain in valid_collection.collection]
subfam_list = [x[0] for x in fam_list]
receptors, counts  = np.unique(fam_list, return_counts=True)
r_counts = list(zip(receptors,counts))
print("HERE IS THE COUNT OF DETECTED aGPCR RECEPTOR PROTEINS:", r_counts, sep="\n")

fam_counts = {}
for prot in fam_list:
    fam = prot[0]
    if fam not in fam_counts.keys():
        fam_counts[fam] = 0
    fam_counts[fam] += 1

print("HERE IS THE COUNT OF DETECTED aGPCR SUBFAMILIES:", fam_counts, sep="\n")

### 3. The receptor groups to be tested are specified. The following sequence of detail was tested:

> - Run1: Subfamily-specific templates
> - Run2: Added Receptor-specific templates for low-matching receptors
> - Run3: Added further templates for extra 14th strand S4

Every subselection is clustered via a pairwise RMSD-matrix constructed from individual _GESAMT_ runs. Two template candidates are then evaluated: The _overall_ centroid and the _largest cluster_ centroid. If they do not match, the clusters are evaluated to see whether there is are separate different clusters. A manual selection of the template is then carried out.

For template selection, we only use 400 randomly selected structures from the respective subset. The templates are tested afterwards to asses their quality. We assume sufficient coverage from these 400 proteins.

# This Box is for running Subselections for Subdomain A >SDA<

best_structures = {}
best_clusters = {}

manual_receptors = ["X"]

run_prefix = "p1"
print(allpdbs)
gain_subset = valid_collection.collection
print(len(gain_subset))

gain_idx_list = range(len(fam_list)) # fam_list
maxlen = 400
data_length = len(gain_subset)
if data_length > maxlen:
    data_length = maxlen # data_length will be used.
    stride = data_length // maxlen
    gain_subset = gain_subset[::stride]
    gain_subset = gain_subset[:maxlen]
    print(len(gain_idx_list), len(gain_subset))

gesamt_outfolder = f"../../TESTING/{run_prefix}_gesamt_sda_pc"

gaingrn.utils.template_utils.run_gesamt_execution(gain_subset, 
                        outfolder=gesamt_outfolder, 
                        pdb_folder=PDB_DIR, 
                        domain='sda', 
                        n_threads=4, 
                        max_struc=maxlen, 
                        gesamt_bin=GESAMT_BIN,
                        no_run=False)

distances = gaingrn.utils.template_utils.evaluate_gesamt_files(gesamt_outfolder, n_prot=data_length, penalty_value=6.0, remove=False)

print(distances.shape)
results = gaingrn.utils.template_utils.cluster_agglomerative(distances, gain_subset, n_cluster=9)
gaingrn.utils.plotting_utils.plot_heirarchy(results['reordered_distances'], groupname=f'aGPCR GAIN', savename=f'{run_prefix}_A_sda_heirarchy.png')
gaingrn.utils.plotting_utils.plot_matrix(results['reordered_distances'], title=f'aGPCR GAIN', savename=f'{run_prefix}_sda_ordered_matrix.png')

all_best = gain_subset[results['overall_best_gain']].name
best_list = []
for r in "ABCDEFGLVX":
    for i, c in results['cluster_best_gains']:
        best_list.append((gain_subset[i].name, c, results['cluster_sizes'][c]))
    best_structures[r] = all_best
    best_clusters[r] = best_list
    print("Done with run",r)

# This Box is for running Subselections for Subdomain B >SDB<
best_structures = {}
best_clusters = {}

for r in 'ABCDEFGLVX': #enumerate(receptors[:1])
    print(r)
    gain_subset = [ gain for i, gain in enumerate(valid_collection.collection) if fam_list[i]==r ]#fam_list[i] == r]#
    print(len(gain_subset))
    gain_idx_list = [ i for i,gain in enumerate(fam_list) if gain == r ] # fam_list
    maxlen = 400
    data_length = len(gain_subset)
    if data_length > maxlen:
        data_length = maxlen # data_length will be used.
        stride = data_length // maxlen
        gain_subset = gain_subset[::stride]
        gain_subset = gain_subset[:maxlen]
        print(len(gain_idx_list), len(gain_subset))

    gesamt_outfolder = f"../{run_prefix}_gesamt_sdb_pc"
    
    gaingrn.utils.template_utils.run_gesamt_execution(gain_subset, outfolder=gesamt_outfolder, gesamt_bin=GESAMT_BIN, pdb_folder=PDB_DIR, domain='sdb', n_threads=4, max_struc=maxlen, no_run=False)
    distances = gaingrn.utils.template_utils.evaluate_gesamt_files(gesamt_outfolder, n_prot=data_length, penalty_value=6.0, remove=False)
    print(distances.shape)
    results = gaingrn.utils.template_utils.cluster_agglomerative(distances, gain_subset, n_cluster=9)
    gaingrn.utils.plotting_utils.plot_heirarchy(results['reordered_distances'], groupname=f'aGPCR GAIN', savename=f'{run_prefix}_sdb_heirarchy.png')
    gaingrn.utils.plotting_utils.plot_matrix(results['reordered_distances'], title=f'aGPCR GAIN', savename=f'{run_prefix}_sdb_ordered_matrix.png')

    all_best = gain_subset[results['overall_best_gain']].name
    best_list = []
    for i, c in results['cluster_best_gains']:
        best_list.append((gain_subset[i].name, c, results['cluster_sizes'][c]))
    best_structures[r] = all_best
    best_clusters[r] = best_list
    print("Done with run",r)

print(best_structures)
print(best_clusters)


#### After creating a set of potential templates, they will each be evaluated against the whole set to remove arbitrarily similar templates, since not every receptor needs a unique template if they are too similar.

#print(results.keys())
#print(results['overall_best_gain'], results['cluster_best_gains'])
all_best = gain_subset[results['overall_best_gain']].name
best_list = []
for i, c in results['cluster_best_gains']:
    best_list.append((gain_subset[i].name, c, results['cluster_sizes'][c]))
print(all_best)
print(best_list)

After the directories with the potential templates have been created, we proceed to __template_testing.ipynb__ to evaluate them and select our final set.
